Fix `submitting` state never getting reset
fixes #487

Previously, when it was checking in the subscribe func if state updated https://github.com/final-form/react-final-form/blob/4ca1ce7a7800c08b249219e14740ec2c1ed7e222/src/ReactFinalForm.js#L89 - the current state `state` - was always the initial state, bec of js closures.

So, the `submitting` got set to `true` bec that was diff than the initial state but when attempting to set back to false.. it was checking original state of `false` with current state of `false` and it was equal and therefore did not update and reset state.

Changed to add `state` to the `useEffect` deps so it will rerun the effect, causing it to create a new subscribe func with the updated state.